update setup script and readme, add github workflow

+++ setup.py
@@ -1,13 +1,20 @@
+import platform
 import subprocess
 import shutil
+import os
+
+WINDOWS: str = "windows"
+LINUX: str = "linux"
+OSX: str = "osx"
 
 
 def setup(args: dict) -> None:
     print("Setting up mouse-tracker")
 
-    print("Building release library")
+    print("Building release library for OS {}".format(args.os), flush=True)
     subprocess.run(["cargo", "build", "--release", "--lib",
                    "--manifest-path=rust/Cargo.toml"], check=True)
+    print("Finished building release library", flush=True)
 
     print("Copying library")
     if args.os == "windows":
@@ -23,7 +30,9 @@ def setup(args: dict) -> None:
 
     if args.export:
         shutil.rmtree("rust")
-        shutil.rmtree("__pycache__")
+        if os.path.isdir("__pycache__"):
+            shutil.rmtree("__pycache__")
+        shutil.rmtree(".github")
 
     print("Finished setting up mouse-tracker")
 
@@ -31,13 +40,17 @@ def setup(args: dict) -> None:
 def clean(_args: dict) -> None:
     print("Cleaning up mouse-tracker")
 
+    print("Running cargo clean", flush=True)
     subprocess.run(
         ["cargo", "clean", "--manifest-path=rust/Cargo.toml"], check=True)
+    print("Finished running cargo clean", flush=True)
 
     print("Finished cleaning up mouse-tracker")
 
 
 if __name__ == "__main__":
+    os.chdir(os.path.dirname(os.path.realpath(__file__)))
+
     from argparse import ArgumentParser
 
     parser = ArgumentParser()
@@ -46,12 +59,24 @@ if __name__ == "__main__":
 
     setup_parser = subparsers.add_parser("setup")
     setup_parser.add_argument(
-        "--os", choices=["windows", "linux", "osx"], required=True)
+        "--os", choices=["windows", "linux", "osx"], default="")
+    setup_parser.add_argument("--export", action="store_true")
     setup_parser.set_defaults(func=setup)
 
     clean_parser = subparsers.add_parser("clean")
     clean_parser.set_defaults(func=clean)
 
     args = parser.parse_args()
+
+    if not args.os:
+        args.os = platform.system()
+        if args.os == "Windows":
+            args.os = WINDOWS
+        elif args.os == "Linux":
+            args.os = LINUX
+        elif args.os == "Darwin":
+            args.os = OSX
+        else:
+            raise Exception("Unhandled OS: {}".format(args.os))
 
     args.func(args)
